requirements.txt
fastapi
uvicorn
numpy
opencv-python
mediapipe
streamlit
Pillow
python-multipart
